+++ R/tidy_rnorm.R
@@ -8,7 +8,10 @@
 #'
 #' @description
 #'
-#' @param
+#' @param .n The number of randomly generated points you want.
+#' @param .mean The mean of the randomly generated data.
+#' @param .sd The standard deviation of the randomly generated data.
+#' @param .num_walks The number of randomly generated simulations you want.
 #'
 #' @examples
 #'
@@ -17,4 +20,6 @@
 #' @export
 #'
 
-tidy_rnorm <- function(.n = 50, .mean = 0, .sd = 0, .num_walks = 1)
+tidy_rnorm <- function(.n = 50, .mean = 0, .sd = 0, .num_walks = 1){
+    return(.n)
+}