+++ NEWS.md
@@ -1,3 +1,5 @@
 # TidyDensity 0.0.0.9000
 
+1. Fix #1 - Add function `tidy_rnorm()`
+
 * Added a `NEWS.md` file to track changes to the package.